Correct ball movement

+++ Popcorn/Level.h
@@ -18,8 +18,9 @@ public:
 
    void Init();
    void Draw(HDC hdc,RECT &paint_area);
-   void Check_Level_Brick_Hit(int &next_y_pos, double &ball_direction);
+   void Check_Level_Brick_Hit(double &next_y_pos, double &ball_direction);
 
+   
    AActive_Brick Active_Brick;
 
 private:
